move dict inference test from capability.spec to inference.spec
